meta: Merged more stuff into shared types
