+++ packages/gcdata/src/index.ts
@@ -1,7 +1,8 @@
-export * from './GameChanger.js';
 export * from './cl2.quest.js';
+export * from './cl2.shared.types.js';
 export * from './cl2.storyline.js';
 export * from './cl2.types.auto.js';
+export * from './GameChanger.js';
 export * from './helpers.js';
 export * from './types.cl2.rumpus.js';
 export * from './types.editor.js';
